// CMIS4-FE/src/app/modules/thu-tien-cham-no-module/models/dung-thanh-toan-mtmn.model.ts
export class CN_DUNGTTOAN_MTMN {
    MA_DVIQLY: string;
    MA_KHANG: string;
    NGAY_KPHUC: string;
    NGAY_KPHUC_DATE: Date;
    NGAY_DUNG: string;
    NGAY_DUNG_DATE: Date;
    SO_BBANKT: string;
    CONG_SUATKT: number;
    SAN_LUONG: number;
    SO_TIEN: number;
    TIEN_GTGT: number;
    LY_DO: string;
    LY_DO_KPHUC: string;
    NGAY_TAO: string;
    NGAY_TAO_DATE: Date;
    NGUOI_TAO: string;
    NGAY_SUA: string;
    NGAY_SUA_DATE: Date;
    NGUOI_SUA: string;
    MA_CNANG: string;

    TEN_KHANG: string;
    CONG_SUAT: any;
    TONG_TIEN: number;
    NGAY_KY: any;
    DIA_CHI: any;
    PHAN_LOAI: any;
}